remove task registry and make task not being a subclass of 'SkillComponent' anymore.
