+++ tests/test_aea.py
@@ -306,7 +306,7 @@ class TestInitializeAEAProgrammaticallyFromResourcesDir:
 
         # TODO the previous code caused an error:
         #      _pickle.PicklingError: Can't pickle <class 'tasks.DummyTask'>: import of module 'tasks' failed
-        dummy_task = DummyTask(name="dummy", skill_context=object())
+        dummy_task = DummyTask()
         async_result = self.aea.task_manager.enqueue_task(dummy_task)
         expected_dummy_task = async_result.get(2.0)
         assert expected_dummy_task.nb_execute_called > 0
@@ -408,7 +408,7 @@ class TestInitializeAEAProgrammaticallyBuildResources:
         assert dummy_behaviour is not None
         assert dummy_behaviour.nb_act_called > 0
 
-        dummy_task = DummyTask(name="dummy", skill_context=object())
+        dummy_task = DummyTask()
         async_result = self.aea.task_manager.enqueue_task(dummy_task)
         expected_dummy_task = async_result.get(2.0)
         assert expected_dummy_task.nb_execute_called > 0
